refactor(v010): carve voicegw reconcile out of _legacy
Twelfth per-command carve-out in section 2.

Adds voicegateway/cli/reconcile.py (105 LOC) owning the 'reconcile'
command. Diffs VG's logged costs against a provider's normalized
usage export. Body unchanged from v0.0.5.

The shared _parse_iso_date_arg helper still lives in _legacy.py:
both export-costs and reconcile import it from there. The helper
moves to a shared helpers module in the section-2 finalisation
when _legacy.py no longer hosts any commands.

Drops 'from pathlib import Path' from _legacy.py: reconcile's
Path(provider_usage_file) was the last consumer.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (13 source
files now).
